+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 0.0.7 (????-??-??)
 ------------------
 
-- `idc-generate-regions` tool can take margins into account for its region calculations now
+- `idc-generate-regions` tool can take margins and overlaps (right/bottom) into account for its region calculations now
 - `idc-combine-sub-images` tool now has more details in the exceptions when extraction of groups fail
   and prunes the annotation after the merge as well
 - `meta-sub-images` now outputs a logging message if there are no annotations after transferring